backend/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
marshmallow==2.18.1
mysqlclient==1.4.2.post1
PyJWT==1.7.1
pytest==4.3.0
SQLAlchemy==1.3.0
pytest-mock==1.10.1
